requirements.txt
transformers
sqlparse
accelerate
bitsandbytes
dbt-postgres
psycopg2-binary
ddlgenerator
pandas